% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiMMA.R
\name{construct_subsets}
\alias{construct_subsets}
\title{Construct Subset Data for Outcome j}
\usage{
construct_subsets(y, s, theta, j)
}
\arguments{
\item{y}{A matrix with effect sizes}

\item{s}{A matrix with standard deviations}

\item{theta}{A matrix with random effect}

\item{j}{A scalar for the outcome j}
}
\value{
A matrix with the subset data D_j contains effect size and standard deviations
}
\description{
Creates a subset of data for a specific outcome.
}
