% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiMMA.R
\name{estimating_function}
\alias{estimating_function}
\title{Compute Estimating Functions for Selection Model}
\usage{
estimating_function(param, subset_data, j)
}
\arguments{
\item{param}{A vector of parameters in the selection model}

\item{subset_data}{A matrix with the subset data D_j contains effect size and standard deviations}

\item{j}{index for outcome j}
}
\value{
A matrix with estimating function values
}
\description{
Evaluates the estimating functions for a semiparametric selection model, facilitating the estimation of parameters that mitigate outcome reporting bias in multivariate meta-analysis.
}
