% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiMMA.R
\name{probability_of_observation}
\alias{probability_of_observation}
\title{Probability of Observing Outcome j in Study i}
\usage{
probability_of_observation(s_i, y_i, alpha_j, beta_j, gamma_j)
}
\arguments{
\item{s_i}{A vector of within-study standard deviation for study i}

\item{y_i}{A vector of effect size for study i}

\item{alpha_j}{A scalar of parameter alpha_j in the selection model for outcome j}

\item{beta_j}{A vector of parameter beta_j in the selection model for outcome j}

\item{gamma_j}{A vector of parameter gamma_j in the selection model for outcome j}
}
\value{
A scalar of the probability for observing the outcome j in study i
}
\description{
Calculates observation probability.
}
