% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiMMA.R
\name{SemiMMA}
\alias{SemiMMA}
\title{Semiparametric Multivariate Meta-Analysis}
\usage{
SemiMMA(input_data)
}
\arguments{
\item{data}{A matrixwith effect sizes and standard deviations in alternating columns, one study for each row}
}
\value{
A matrix with estimates of effect sizes and standard deviations for each outcome  per row
}
\description{
Conducts a semiparametric multivariate meta-analysis to estimate effect sizes and standard deviations.
}
